requirements.txt
Flask~=3.1.0
Werkzeug~=3.1.3
python-dotenv~=1.0.1